+++ R/quadratic_forms.R
@@ -65,6 +65,9 @@
 #' with one column for each level of sampling to be accounted for by the variance estimator.
 #' @param sort_order Required if \code{variance_estimator} equals \code{"SD1"} or \code{"SD2"}.
 #' This should be a vector that orders the rows of data into the order used for sampling.
+#' @section Variance Estimators:
+#' See \link[svrep]{variance-estimators} for a
+#' description of each variance estimator.
 #' @section Arguments required for each variance estimator:
 #' Below are the arguments that are required or optional for each variance estimator.
 #'
@@ -78,118 +81,13 @@
 #' | Deville-2                | Required   |             | Required    | Optional    |                  |            |
 #' | Yates-Grundy             |            | Required    |             |             |                  |            |
 #' | Horvitz-Thompson         |            | Required    |             |             |                  |            |
-#' @section Variance Estimators:
-#' The \strong{Horvitz-Thompson} variance estimator:
-#' \deqn{
-#'   v(\hat{Y}) = \sum_{i \in s}\sum_{j \in s} (1 - \frac{\pi_i \pi_j}{\pi_{ij}}) \frac{y_i}{\pi_i} \frac{y_j}{\pi_j}
-#' }
-#' The \strong{Yates-Grundy} variance estimator:
-#' \deqn{
-#'   v(\hat{Y}) = -\frac{1}{2}\sum_{i \in s}\sum_{j \in s} (1 - \frac{\pi_i \pi_j}{\pi_{ij}}) (\frac{y_i}{\pi_i} - \frac{y_j}{\pi_j})^2
-#' }
-#' The \strong{Poisson Horvitz-Thompson} variance estimator
-#' is simply the Horvitz-Thompson variance estimator, but
-#' where \eqn{\pi_{ij}=\pi_i \times \pi_j}, which is the case for Poisson sampling.
-#' \cr \cr
-#' The \strong{Stratified Multistage SRS} variance estimator is the recursive variance estimator
-#' proposed by Bellhouse (1985) and used in the 'survey' package's function \link[survey]{svyrecvar}.
-#' The estimator can be used for any number of sampling stages. For illustration, we describe its use
-#' for two sampling stages.
-#' \deqn{
-#'   v(\hat{Y}) = \hat{V}_1 + \hat{V}_2
-#' }
-#' where
-#' \deqn{
-#'   \hat{V}_1 = \sum_{h=1}^{H} (1 - \frac{n_h}{N_h})\frac{n_h}{n_h - 1} \sum_{i=1}^{n_h} (y_{hi.} - \bar{y}_{hi.})^2
-#' }
-#' and
-#' \deqn{
-#'   \hat{V}_2 = \sum_{h=1}^{H} \frac{n_h}{N_h} \sum_{i=1}^{n_h}v_{hi}(y_{hi.})
-#' }
-#' where \eqn{n_h} is the number of sampled clusters in stratum \eqn{h},
-#' \eqn{N_h} is the number of population clusters in stratum \eqn{h},
-#' \eqn{y_{hi.}} is the weighted cluster total in cluster \eqn{i} of stratum \eqn{h},
-#' \eqn{\bar{y}_{hi.}} is the mean weighted cluster total of stratum \eqn{h},
-#' (\eqn{\bar{y}_{hi.} = \frac{1}{n_h}\sum_{i=1}^{n_h}y_{hi.}}), and
-#' \eqn{v_{hi}(y_{hi.})} is the estimated sampling variance of \eqn{y_{hi.}}.
-#' \cr \cr
-#' The \strong{Ultimate Cluster} variance estimator is simply the stratified multistage SRS
-#' variance estimator, but ignoring variances from later stages of sampling.
-#' \deqn{
-#'   v(\hat{Y}) = \hat{V}_1
-#' }
-#' This is the variance estimator used in the 'survey' package when the user specifies
-#' \code{option(survey.ultimate.cluster = TRUE)} or uses \code{svyrecvar(..., one.stage = TRUE)}.
-#' When the first-stage sampling fractions are small, analysts often omit the finite population corrections \eqn{(1-\frac{n_h}{N_h})}
-#' when using the ultimate cluster estimator.
-#' \cr \cr
-#' The \strong{SD1} and \strong{SD2} variance estimators are "successive difference"
-#' estimators sometimes used for systematic sampling designs.
-#' Ash (2014) describes each estimator as follows:
-#' \deqn{
-#'   \hat{v}_{S D 1}(\hat{Y}) = \left(1-\frac{n}{N}\right) \frac{n}{2(n-1)} \sum_{k=2}^n\left(\breve{y}_k-\breve{y}_{k-1}\right)^2
-#' }
-#' \deqn{
-#'   \hat{v}_{S D 2}(\hat{Y}) = \left(1-\frac{n}{N}\right) \frac{1}{2}\left[\sum_{k=2}^n\left(\breve{y}_k-\breve{y}_{k-1}\right)^2+\left(\breve{y}_n-\breve{y}_1\right)^2\right]
-#' }
-#' where \eqn{\breve{y}_k = y_k/\pi_k} is the weighted value of unit \eqn{k}
-#' with selection probability \eqn{\pi_k}. The SD1 estimator is recommended by Wolter (1984).
-#' The SD2 estimator is the basis of the successive difference replication estimator commonly
-#' used for systematic sampling designs. See Ash (2014) for details.
-#' \cr \cr
-#' For multistage samples, SD1 and SD2 are applied to the clusters at each stage, separately by stratum.
-#' For later stages of sampling, the variance estimate from a stratum is multiplied by the product
-#' of sampling fractions from earlier stages of sampling. For example, at a third stage of sampling,
-#' the variance estimate from a third-stage stratum is multiplied by \eqn{\frac{n_1}{N_1}\frac{n_2}{N_2}},
-#' which is the product of sampling fractions from the first-stage stratum and second-stage stratum.
-#' \cr \cr
-#' The \strong{"Deville-1"} and \strong{"Deville-2"} variance estimators
-#' are clearly described in Matei and Tillé (2005),
-#' and are intended for designs that use
-#' fixed-size, unequal-probability random sampling without replacement.
-#' These variance estimators have been shown to be effective
-#' for designs that use a fixed sample size with a high-entropy sampling method.
-#' This includes most PPSWOR sampling methods,
-#' but unequal-probability systematic sampling is an important exception.
-#'
-#' These variance estimators take the following form:
-#' \deqn{
-#' \hat{v}(\hat{Y}) = \sum_{i=1}^{n} c_i (\breve{y}_i - \frac{1}{\sum_{i=k}^{n}c_k}\sum_{k=1}^{n}c_k \breve{y}_k)^2
-#' }
-#' where \eqn{\breve{y}_i = y_i/\pi_i} is the weighted value of the the variable of interest,
-#' and \eqn{c_i} depend on the method used:
-#' \itemize{
-#'   \item{\strong{"Deville-1"}: }{
-#'     \deqn{c_i=\left(1-\pi_i\right) \frac{n}{n-1}}}
-#'   \item{\strong{"Deville-2"}: }{
-#'     \deqn{c_i = (1-\pi_i) \left[1 - \sum_{k=1}^{n} \left(\frac{1-\pi_k}{\sum_{k=1}^{n}(1-\pi_k)}\right)^2 \right]^{-1}}}
-#' }
-#' In the case of simple random sampling without replacement (SRSWOR),
-#' these estimators are both identical to the usual stratified multistage SRS estimator
-#' (which is itself a special case of the Horvitz-Thompson estimator).
-#'
-#' For multistage samples, "Deville-1" and "Deville-2" are applied to the clusters at each stage, separately by stratum.
-#' For later stages of sampling, the variance estimate from a stratum is multiplied by the product
-#' of sampling probabilities from earlier stages of sampling. For example, at a third stage of sampling,
-#' the variance estimate from a third-stage stratum is multiplied by \eqn{\pi_1 \times \pi_{(2 | 1)}},
-#' where \eqn{\pi_1} is the sampling probability of the first-stage unit
-#' and \eqn{\pi_{(2|1)}} is the sampling probability of the second-stage unit
-#' within the first-stage unit.
-#' @references
-#' Ash, S. (2014). "\emph{Using successive difference replication for estimating variances}."
-#' \strong{Survey Methodology}, Statistics Canada, 40(1), 47–59.
-#'
-#' Bellhouse, D.R. (1985). "\emph{Computing Methods for Variance Estimation in Complex Surveys}."
-#' \strong{Journal of Official Statistics}, Vol.1, No.3.
-#'
-#' Matei, Alina, and Yves Tillé. (2005).
-#' “\emph{Evaluation of Variance Approximations and Estimators
-#' in Maximum Entropy Sampling with Unequal Probability and Fixed Sample Size.}”
-#' \strong{Journal of Official Statistics}, 21(4):543–70.
 #' @md
 #' @return The matrix of the quadratic form representing the variance estimator.
 #' @export
 #' @seealso
+#' See \link[svrep]{variance-estimators} for a
+#' description of each variance estimator.
+#'
 #' For a two-phase design, the function
 #' \link[svrep]{make_twophase_quad_form} combines
 #' the quadratic form matrix from each phase.
@@ -296,11 +194,14 @@ make_quad_form_matrix <- function(variance_estimator = "Yates-Grundy",
       stop("`joint_probs` must be a matrix of values between 0 and 1, with no missing values.")
     }
     number_of_ultimate_units <- ncol(joint_probs)
+    use_sparse_matrix <- FALSE
   }
 
   # Check inputs and assemble all necessary information
   # for estimators of stratified/clustered designs
   if (variance_estimator %in% c("Stratified Multistage SRS", "Ultimate Cluster", "SD1", "SD2", "Deville-1", "Deville-2")) {
+
+    use_sparse_matrix <- TRUE
 
     # Ensure the minimal set of inputs is supplied
     if (variance_estimator %in% c("Stratified Multistage SRS", "Ultimate Cluster", "Deville-1", "Deville-2")) {
@@ -347,14 +248,14 @@ make_quad_form_matrix <- function(variance_estimator = "Yates-Grundy",
     if (is.null(cluster_ids) && !is.null(strata_ids)) {
       number_of_stages <- ncol(strata_ids)
       number_of_ultimate_units <- nrow(strata_ids)
-      cluster_ids <- matrix(data = rep(seq_len(number_of_ultimate_units),
+      cluster_ids <- Matrix::Matrix(data = rep(seq_len(number_of_ultimate_units),
                                        times = number_of_stages),
                             nrow = number_of_ultimate_units, ncol = number_of_stages,
                             byrow = FALSE)
     } else if (!is.null(cluster_ids) && is.null(strata_ids)) {
       number_of_stages <- ncol(cluster_ids)
       number_of_ultimate_units <- nrow(cluster_ids)
-      strata_ids <- matrix(1,
+      strata_ids <- Matrix::Matrix(1,
                            nrow = number_of_ultimate_units,
                            ncol = number_of_stages)
     } else {
@@ -385,9 +286,8 @@ make_quad_form_matrix <- function(variance_estimator = "Yates-Grundy",
       }
     }
 
-
     if (is.null(strata_pop_sizes)) {
-      strata_pop_sizes <- matrix(data = Inf,
+      strata_pop_sizes <- Matrix::Matrix(data = Inf,
                                  nrow = number_of_ultimate_units,
                                  ncol = number_of_stages)
     }
@@ -395,7 +295,7 @@ make_quad_form_matrix <- function(variance_estimator = "Yates-Grundy",
 
   if (variance_estimator == "Horvitz-Thompson") {
     n <- number_of_ultimate_units
-    quad_form_matrix <- matrix(nrow = n, ncol = n)
+    quad_form_matrix <- Matrix::Matrix(nrow = n, ncol = n) |> as("symmetricMatrix")
     for (i in seq_len(n)) {
       for (j in seq(from = i, to = n, by = 1)) {
         quad_form_matrix[i,j] <- 1 - (joint_probs[i,i] * joint_probs[j,j])/joint_probs[i,j]
@@ -406,21 +306,25 @@ make_quad_form_matrix <- function(variance_estimator = "Yates-Grundy",
 
   if (variance_estimator == "Yates-Grundy") {
     n <- number_of_ultimate_units
-    quad_form_matrix <- matrix(nrow = n, ncol = n)
+    quad_form_matrix <- Matrix::Matrix(nrow = n, ncol = n) |> as("symmetricMatrix")
     for (i in seq_len(n)) {
       for (j in seq(from = i, to = n, by = 1)) {
         quad_form_matrix[i,j] <- -(1 - (joint_probs[i,i] * joint_probs[j,j])/joint_probs[i,j])
       }
     }
     quad_form_matrix[lower.tri(quad_form_matrix)] <- t(quad_form_matrix)[lower.tri(quad_form_matrix)]
-    diag(quad_form_matrix) <- diag(quad_form_matrix) - rowSums(quad_form_matrix)
+    diag(quad_form_matrix) <- Matrix::diag(quad_form_matrix) - Matrix::rowSums(quad_form_matrix)
     quad_form_matrix <- -1 * quad_form_matrix
   }
 
   if (variance_estimator == "Ultimate Cluster") {
-    quad_form_matrix <- matrix(data = 0,
+    quad_form_matrix <- Matrix::Matrix(data = 0,
                                nrow = number_of_ultimate_units,
-                               ncol = number_of_ultimate_units)
+                               ncol = number_of_ultimate_units) |>
+      as("symmetricMatrix")
+    if (use_sparse_matrix) {
+      quad_form_matrix <- quad_form_matrix |> as("CsparseMatrix")
+    }
 
     # Generate quadratic form for each stratum
     for (stratum_id in unique(strata_ids[,1,drop=TRUE])) {
@@ -437,9 +341,15 @@ make_quad_form_matrix <- function(variance_estimator = "Yates-Grundy",
   }
 
   if (variance_estimator %in% c("Stratified Multistage SRS", "Deville-1", "Deville-2")) {
-    quad_form_matrix <- matrix(data = 0,
-                               nrow = number_of_ultimate_units,
-                               ncol = number_of_ultimate_units)
+    quad_form_matrix <- Matrix::Matrix(data = 0,
+                                       nrow = number_of_ultimate_units,
+                                       ncol = number_of_ultimate_units) |>
+      as("symmetricMatrix")
+
+    if (use_sparse_matrix) {
+      quad_form_matrix <- quad_form_matrix |> as("CsparseMatrix")
+    }
+
     # Obtain matrix of cluster sample sizes by stage
     count <- function(x) sum(!duplicated(x))
     sampsize <- matrix(ncol = ncol(cluster_ids), nrow = nrow(cluster_ids))
@@ -519,7 +429,11 @@ make_quad_form_matrix <- function(variance_estimator = "Yates-Grundy",
   if (variance_estimator %in% c("SD1", "SD2")) {
     n <- number_of_ultimate_units
     # Initialize quadratic form matrix
-    quad_form_matrix <- matrix(data = 0, nrow = n, ncol = n)
+    quad_form_matrix <- Matrix::Matrix(data = 0, nrow = n, ncol = n) |>
+      as("symmetricMatrix")
+    if (use_sparse_matrix) {
+      quad_form_matrix <- quad_form_matrix |> as("CsparseMatrix")
+    }
     sorted_quad_form_matrix <- quad_form_matrix
     # Sort the inputs, compile into a dataframe
     sorted_df <- data.frame('Row_ID' = seq_len(n),
@@ -590,10 +504,10 @@ make_sd_matrix <- function(n, f = 0, type = "SD1") {
   }
 
   if (n == 1) {
-    C_matrix <- matrix(0, nrow = 1, ncol = 1)
+    C_matrix <- Matrix::Matrix(0, nrow = 1, ncol = 1)
   }
   if (type == "SD1" && (n > 1)) {
-    C_matrix <- diag(n) * 2
+    C_matrix <- as(Matrix::Diagonal(n) * 2, "symmetricMatrix")
     C_matrix[1,1] <- 1
     C_matrix[n,n] <- 1
     for (i in seq_len(n)) {
@@ -601,7 +515,7 @@ make_sd_matrix <- function(n, f = 0, type = "SD1") {
         C_matrix[i,i+1] <- -1
         C_matrix[i+1,i] <- -1
       }
-      if (i > 0) {
+      if (i > 1) {
         C_matrix[i,i-1] <- -1
         C_matrix[i-1,i] <- -1
       }
@@ -609,13 +523,13 @@ make_sd_matrix <- function(n, f = 0, type = "SD1") {
     C_matrix <- (n/(n-1)) * C_matrix
   }
   if (type == "SD2" && (n > 1)) {
-    C_matrix <- diag(n) * 2
+    C_matrix <- as(Matrix::Diagonal(n) * 2, "symmetricMatrix")
     for (i in seq_len(n)) {
       if (i < n) {
         C_matrix[i,i+1] <- -1
         C_matrix[i+1,i] <- -1
       }
-      if (i > 0) {
+      if (i > 1) {
         C_matrix[i,i-1] <- -1
         C_matrix[i-1,i] <- -1
       }
@@ -637,7 +551,7 @@ make_sd_matrix <- function(n, f = 0, type = "SD1") {
 
 #' @title Create a quadratic form's matrix to represent the basic variance estimator
 #' for a total under simple random sampling without replacement
-#' @description The usual variance estimator simple random sampling without replacement
+#' @description The usual variance estimator for simple random sampling without replacement
 #' can be represented as a quadratic form.
 #' This function determines the matrix of the quadratic form.
 #' @param n Sample size
@@ -664,9 +578,9 @@ make_srswor_matrix <- function(n, f = 0) {
     stop("`f` must be a single number between 0 and 1")
   }
   if (n == 1) {
-    C_matrix <- matrix(0, nrow = 1, ncol = 1)
+    C_matrix <- Matrix::Matrix(0, nrow = 1, ncol = 1)
   } else {
-    C_matrix <- matrix(-1/(n-1), nrow = n, ncol = n)
+    C_matrix <- Matrix::Matrix(-1/(n-1), nrow = n, ncol = n)
     diag(C_matrix) <- 1
     C_matrix <- (1-f) * C_matrix
   }
@@ -711,6 +625,9 @@ make_srswor_matrix <- function(n, f = 0) {
 #'   \sigma_{ii} = c_i (1 - \frac{c_i}{\sum_{k=1}^{n}c_k}) \textit{ when } i = j \\
 #'   \sigma_{ij}=\frac{-c_i c_j}{\sum_{k=1}^{n}c_k} \textit{ when } i \neq j \\
 #' }
+#' When \eqn{\pi_{i} = 1} for every unit, then \eqn{\sigma_{ij}=0} for all \eqn{i,j}.
+#' If there is only one sampling unit, then \eqn{\sigma_{11}=0}; that is, the unit is treated as if it was sampled with certainty.
+#'
 #' The constants \eqn{c_i} are defined for each approximation method as follows,
 #' with the names taken directly from Matei and Tillé (2005).
 #' \itemize{
@@ -735,10 +652,10 @@ make_srswor_matrix <- function(n, f = 0) {
 #' @keywords internal
 make_ppswor_approx_matrix <- function(probs, method = "Deville-1") {
 
+  n <- length(probs)
   one_minus_pi <- 1 - probs
 
   if (method == "Deville-1") {
-    n <- length(probs)
     c_k <- (1 - probs) * (n/(n-1))
   }
   if (method == "Deville-2") {
@@ -749,8 +666,13 @@ make_ppswor_approx_matrix <- function(probs, method = "Deville-1") {
 
   c_sum <- sum(c_k)
 
-  Sigma <- outer(c_k, -c_k) / c_sum
-  diag(Sigma) <- c_k*(1 - c_k/c_sum)
+  if ((n == 1) || (c_sum == 0)) {
+    Sigma <- Matrix::Matrix(0, nrow = length(c_k), ncol = length(c_k))
+  } else {
+    Sigma <- outer(c_k, -c_k) / c_sum
+    diag(Sigma) <- c_k*(1 - c_k/c_sum)
+    Sigma <- as(Sigma, "symmetricMatrix")
+  }
 
   return(Sigma)
 }
@@ -837,7 +759,12 @@ distribute_matrix_across_clusters <- function(cluster_level_matrix, cluster_ids,
 #' eigen(X)$values
 #' @export
 is_psd_matrix <- function(X, tolerance = sqrt(.Machine$double.eps)) {
-  symmetric <- isSymmetric(X)
+  if (inherits(X, 'symmetricMatrix')) {
+    symmetric <- TRUE
+  } else {
+    symmetric <- Matrix::isSymmetric(X)
+  }
+
   if (!symmetric) {
     result <- FALSE
   } else {
@@ -908,7 +835,7 @@ get_nearest_psd_matrix <- function(X) {
 #' @return The matrix of joint inclusion probabilities
 #' @keywords internal
 ht_matrix_to_joint_probs <- function(ht_quad_form) {
-  first_order_probs <- 1 - diag(ht_quad_form)
+  first_order_probs <- 1 - Matrix::diag(ht_quad_form)
   joint_probs <- ((1 - ht_quad_form)^(-1)) * outer(first_order_probs, first_order_probs)
   return(joint_probs)
 }
@@ -1079,7 +1006,7 @@ ht_matrix_to_joint_probs <- function(ht_quad_form) {
 #'   n <- twophase_design$phase2$fpc$sampsize[1,1]
 #'   N <- twophase_design$phase2$fpc$popsize[1,1]
 #'
-#'   second_phase_joint_probs <- matrix((n/N)*((n-1)/(N-1)),
+#'   second_phase_joint_probs <- Matrix::Matrix((n/N)*((n-1)/(N-1)),
 #'                                      nrow = n, ncol = n)
 #'   diag(second_phase_joint_probs) <- rep(n/N, times = n)
 #'
@@ -1198,7 +1125,7 @@ make_twophase_quad_form <- function(sigma_1, sigma_2, phase_2_joint_probs,
                                     ensure_psd = TRUE) {
 
   # Diagonal matrix whose entries are second-phase first-order probabilities
-  phase2_prob_matrix <- diag(diag(phase_2_joint_probs))
+  phase2_prob_matrix <- Matrix::diag(Matrix::diag(phase_2_joint_probs))
 
   # Weighted version of `Sigma_1`
   wtd_sigma_1 <- sigma_1 / phase_2_joint_probs