+++ tests/testthat/test-as_bootstrap_design.R
@@ -4,6 +4,7 @@ suppressWarnings({
     library(dplyr)
     library(svrep)
     library(testthat)
+    library(srvyr)
   })
 })
 
@@ -121,11 +122,23 @@ set.seed(2014)
 
   test_that(
     desc = "Returns `tbl_svy` if the input is a `tbl_svy` and 'srvyr' is loaded", {
-      library(srvyr)
       expect_true(
         dstrat |> as_survey() |>
           as_bootstrap_design(replicates = 1) |>
           inherits(what = "tbl_svy")
       )
+    }
+  )
+
+# Informative error message ----
+
+  test_that(
+    desc = "Informative error message if user misspecifies `type` argument.", {
+
+      expect_error(regexp = "Must use either",
+        dstrat |> as_bootstrap_design(replicates = 1,
+                                      type = "bootsy-collins")
+      )
+
     }
   )